requirements.txt
streamlit
google-generativeai
python-dotenv
pymupdf
langchain
PyPDF2
faiss-cpu
langchain_google_genai
langchain-community